Add justify content property
Make sure we are mimicking float behaviour when image is on the right. If we don't add space between then the image sits directly next to the content. For example this doesn't work if no width on content and and there's only a short paragraph which doesn't reach full width of parent container. Want layout to be predictable in this respect.